rename: 컴포넌트 PageTitle, PageContainer는 IntroPage에서 주로 사용되어 파일 위치를 IntroLayout 하위로 이동 (#90)
