+++ src/components/index.ts
@@ -4,7 +4,7 @@ export { default as FileUploading } from './FileUploading';
 export { default as Layout } from './Layout';
 export { default as Loading } from './Loading';
 export { default as Pagination } from './Pagination';
-export { default as PageTitle } from './PageTitle';
-export { default as PageContainer } from './PageContainer';
+export { default as PageTitle } from './IntroLayout/PageTitle';
+export { default as PageContainer } from './IntroLayout/PageContainer';
 export { default as IntroLayout } from './IntroLayout';
 export { default as UserProfilePhoto } from './UserProfilePhoto';
